+++ engines/mmap.h
@@ -12,7 +12,7 @@ void mmap_rand_write(Mapping *, Results *, Arguments *);
 void mmap_prepare_mapping(Mapping *, Arguments);
 void mmap_cleanup_mapping(Mapping *);
 void mmap_run_benchmark(Mapping *mapping, Arguments *args, Results *results);
-void mmap_prepare_map_anon(Mapping *mapping, int fsize);
+void mmap_prepare_map_anon(Mapping *mapping, uint64_t fsize);
 void mmap_init_file(int, int, int);
 void mmap_init_mem(Mapping *);
 #endif
